pyproject.toml
[project]
name = "indic-conformer-asr-testing"
version = "0.1.0"
description = "Indic Conformer ASR Testing API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "transformers",
    "torch",
    "torchaudio",
    "python-multipart",
    "librosa",
    "soundfile",
    "numpy",
    "python-dotenv",
    "onnx",
    "onnxruntime-gpu",
    "datasets",
    "jiwer",
    "pandas",
    "pipecat-ai[local]>=0.0.36",
    "websockets>=15.0.1",
    "loguru>=0.7.0",
]

[tool.uv]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }
